Show default size-cutoff value in --help output
It wasn't clear to me from the current `--help` output that the
default `--size-cutoff` value is 1000 bytes. This patch tweaks the
`--help` output to make this explicit. This makes use of
`optparse-applicative`'s `showDefaultWith` modifier, which was
introduced in version 0.3.1, so this patch also bumps the lower
version bounds on `optparse-applicative` accordingly.