Add support for bazel
This commit is super specific to bazel environment.

In Bazel, when you run a test with `bazel test`, they are run in a
sandbox, so files written by the golden test `--accept` will only be
modified in the sandbox, but not in your original source tree.

This commit detects that you are using bazel and change the output
directory to the original directory in your source tree.